+++ packages/lib/zuauth/src/index.ts
@@ -1,136 +1,9 @@
-import { EdDSATicketPCDTypeName } from "@pcd/eddsa-ticket-pcd/EdDSATicketPCD";
-import type { PipelineEdDSATicketPCDMetadata } from "@pcd/passport-interface";
-import { constructZupassPcdGetRequestUrl } from "@pcd/passport-interface/PassportInterface";
 import {
-  PopupActionResult,
-  zupassPopupExecute,
+  ZuAuthArgs,
+  zuAuthPopup,
+  zuAuthRedirect,
   zupassPopupSetup
-} from "@pcd/passport-interface/PassportPopup/core";
-import { ArgumentTypeName } from "@pcd/pcd-types";
-import { SemaphoreIdentityPCDTypeName } from "@pcd/semaphore-identity-pcd/SemaphoreIdentityPCD";
-import { ZKEdDSAEventTicketPCDPackage } from "@pcd/zk-eddsa-event-ticket-pcd";
-import {
-  EdDSATicketFieldsToReveal,
-  ZKEdDSAEventTicketPCDArgs,
-  ZKEdDSAEventTicketPCDTypeName
-} from "@pcd/zk-eddsa-event-ticket-pcd/ZKEdDSAEventTicketPCD";
+} from "./zuauth";
 
-// For convenience, re-export this here so that consumers don't need to install
-// and import from `@pcd/passport-interface` directly.
-export { zupassPopupSetup };
-
-/**
- * Arguments required for making ZK proofs about tickets.
- */
-export interface ZuAuthArgs {
-  zupassUrl?: string;
-  returnUrl: string;
-  fieldsToReveal: EdDSATicketFieldsToReveal;
-  watermark: string | bigint;
-  // Event metadata comes from Podbox, and identifies the public key, event ID,
-  // and product IDs that are used for issuance by a given pipeline.
-  eventTicketMetadata: PipelineEdDSATicketPCDMetadata[];
-  externalNullifier?: string | bigint;
-  proofTitle?: string;
-  proofDescription?: string;
-}
-
-/**
- * Opens a popup window to the Zupass prove screen.
- */
-export async function zuAuthPopup(
-  args: ZuAuthArgs
-): Promise<PopupActionResult> {
-  const proofUrl = constructZkTicketProofUrl(args);
-  return zupassPopupExecute(args.returnUrl, proofUrl);
-}
-
-/**
- * Navigates to the Zupass prove screen.
- */
-export function zuAuthRedirect(args: ZuAuthArgs): void {
-  const proofUrl = constructZkTicketProofUrl(args);
-  window.location.href = proofUrl;
-}
-
-/**
- * Constructs a URL to the Zupass prove screen for a ZKEdDSAEventTicketPCD
- * zero-knowlege proof.
- */
-export function constructZkTicketProofUrl(zuAuthArgs: ZuAuthArgs): string {
-  const {
-    zupassUrl = "https://zupass.org",
-    returnUrl,
-    fieldsToReveal,
-    watermark,
-    eventTicketMetadata,
-    externalNullifier,
-    proofTitle = "ZKEdDSA Ticket Proof",
-    proofDescription = "ZKEdDSA Ticket PCD Request"
-  } = zuAuthArgs;
-
-  const eventIds = [],
-    productIds = [],
-    publicKeys = [];
-
-  for (const em of eventTicketMetadata) {
-    eventIds.push(em.eventId);
-    productIds.push(em.productId);
-    publicKeys.push(em.publicKey);
-  }
-
-  const args: ZKEdDSAEventTicketPCDArgs = {
-    ticket: {
-      argumentType: ArgumentTypeName.PCD,
-      pcdType: EdDSATicketPCDTypeName,
-      value: undefined,
-      userProvided: true,
-      validatorParams: {
-        eventIds,
-        productIds,
-        publicKeys,
-        notFoundMessage: "No eligible PCDs found"
-      }
-    },
-    identity: {
-      argumentType: ArgumentTypeName.PCD,
-      pcdType: SemaphoreIdentityPCDTypeName,
-      value: undefined,
-      userProvided: true
-    },
-    validEventIds: {
-      argumentType: ArgumentTypeName.StringArray,
-      value:
-        eventIds.length !== 0 && eventIds.length <= 20 ? eventIds : undefined,
-      userProvided: false
-    },
-    fieldsToReveal: {
-      argumentType: ArgumentTypeName.ToggleList,
-      value: fieldsToReveal,
-      userProvided: false
-    },
-    watermark: {
-      argumentType: ArgumentTypeName.BigInt,
-      value: watermark.toString(),
-      userProvided: false
-    },
-    externalNullifier: {
-      argumentType: ArgumentTypeName.BigInt,
-      value: externalNullifier
-        ? externalNullifier.toString()
-        : watermark.toString(),
-      userProvided: false
-    }
-  };
-  return constructZupassPcdGetRequestUrl<typeof ZKEdDSAEventTicketPCDPackage>(
-    zupassUrl,
-    returnUrl,
-    ZKEdDSAEventTicketPCDTypeName,
-    args,
-    {
-      genericProveScreen: true,
-      title: proofTitle,
-      description: proofDescription
-    }
-  );
-}
+export { zuAuthPopup, zuAuthRedirect, zupassPopupSetup };
+export type { ZuAuthArgs };
